implemented few repositories, TourPage doesn't work database prob

+++ src/main/java/com/example/marketour/model/audio/Audio.java
@@ -13,7 +13,7 @@ public class Audio {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "audio_id")
-    private int audioId;
+    private Long audioId;
 
     @Column(name = "description")
     private String description;
